marine raider snowflake

+++ code/game/objects/items/devices/helmet_visors.dm
@@ -211,6 +211,9 @@
 	/// A slight glowing green light while the NVG is activated, is initialized as in the attached_helmet's contents
 	var/atom/movable/nvg_light/on_light
 
+	/// Whether or not the sight uses on_light and produces light
+	var/visor_glows = TRUE
+
 /obj/item/device/helmet_visor/night_vision/Initialize(mapload, ...)
 	. = ..()
 	power_cell = new(src)
@@ -231,8 +234,9 @@
 	user.overlay_fullscreen("nvg_visor", /atom/movable/screen/fullscreen/flash/noise/nvg)
 	user.overlay_fullscreen("nvg_visor_blur", /atom/movable/screen/fullscreen/brute/nvg, 3)
 	user.update_sight()
-	on_light = new(attached_helmet)
-	on_light.set_light_on(TRUE)
+	if(visor_glows)
+		on_light = new(attached_helmet)
+		on_light.set_light_on(TRUE)
 	START_PROCESSING(SSobj, src)
 
 /obj/item/device/helmet_visor/night_vision/deactivate_visor(obj/item/clothing/head/helmet/marine/attached_helmet, mob/living/carbon/human/user)
@@ -240,7 +244,8 @@
 	user.clear_fullscreen("nvg_visor", 0.5 SECONDS)
 	user.clear_fullscreen("nvg_visor_blur", 0.5 SECONDS)
 
-	qdel(on_light)
+	if(visor_glows)
+		qdel(on_light)
 	UnregisterSignal(user, COMSIG_HUMAN_POST_UPDATE_SIGHT)
 
 	user.update_sight()
@@ -298,7 +303,9 @@
 	name = "advanced night vision optic"
 	desc = "An insertable visor HUD into a standard USCM helmet. This type gives a form of night vision and is standard issue in special forces units."
 	hud_type = list(MOB_HUD_FACTION_USCM, MOB_HUD_MEDICAL_ADVANCED)
+	helmet_overlay = "nvg_sight_right_raider"
 	power_use = 0
+	visor_glows = FALSE
 
 /obj/item/device/helmet_visor/night_vision/marine_raider/activate_visor(obj/item/clothing/head/helmet/marine/attached_helmet, mob/living/carbon/human/user)
 	. = ..()
